pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "neural-unsupervised-electrode-mapping"
version = "0.1.0"
description = "Minimal receptive field reconstruction demo"
authors = [{ name = "neural-unsupervised-electrode-mapping" }]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "numpy>=1.24",
    "scipy>=1.11",
    "scikit-learn>=1.3",
    "umap-learn>=0.5",
    "matplotlib>=3.7",
    "seaborn>=0.12",
    "pandas>=2.0",
    "pytest>=7.4",
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["rf_demo", "rf_demo.*"]

